new report page.
I started the report page but I forgot what to put for report no2 . I
gave an example for the character report model. if you could send me a
message with details about report I would like it. Also a comment about
it even if its not done would be nice